update with comments and use ts
