+++ package.json
@@ -1,11 +1,47 @@
 {
   "name": "indigo-tokens",
-  "version": "1.2.3",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+  "version": "1.2.4",
+  "description": "Tlon&#x27;s design language",
+  "author": "tlon",
+  "license": "MIT",
+  "repository": "urbit/indigo-tokens",
+  "main": "dist/index.js",
+  "module": "dist/indigo-react.esm.js",
+  "files": [
+    "dist"
+  ],
+  "engines": {
+    "node": ">=8",
+    "npm": ">=5"
   },
-  "author": "",
-  "license": "ISC"
+  "scripts": {
+    "start": "tsdx watch",
+    "build": "tsdx build",
+    "test": "tsdx test --passWithNoTests",
+    "lint": "tsdx lint",
+    "lint:fix": "tsdx lint src --fix",
+    "prepare": "tsdx build"
+  },
+  "peerDependencies": {
+    "styled-components": ">= 4"
+  },
+  "devDependencies": {
+    "@types/styled-components": "^4.4.3",
+    "husky": "^4.2.2",
+    "styled-components": "^5.0.1",
+    "tsdx": "^0.12.3",
+    "tslib": "^1.10.0",
+    "typescript": "^3.7.5"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "tsdx lint"
+    }
+  },
+  "prettier": {
+    "printWidth": 80,
+    "semi": true,
+    "singleQuote": true,
+    "trailingComma": "es5"
+  }
 }
